Add --max-log-lines option to set # of items to print per section
Add an optional CLI option --max-log-lines for setting the number of
lines of build info to print for each section. All other lines in that
section will be hidden and "+ X more" will be displayed instead. This
prevents the log from becoming too long when working on apps with
thousands of pages.

The option defaults to showing 50 lines per build section: 50 lines for
SSR pages, 50 lines for SSG pages, 50 lines of redirects, ...

See: https://github.com/FinnWoelm/next-on-netlify/issues/26